# api/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Documentação da API de delivery Cardapiogo.",
        "title": "Cardapiogo API",
        "contact": {
            "name": "Gustavo Ferreira de Jesus",
            "email": "gustavo081900@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
        "/v1/auth/refresh": {
            "post": {
                "description": "Refresh an expired access token by providing a valid refresh token. Returns a new JWT if the refresh is successful.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh JWT using a refresh token",
                "parameters": [
                    {
                        "description": "Request body containing the refresh token.",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authentication.RefreshParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a new JWT upon successful token refresh.",
                        "schema": {
                            "$ref": "#/definitions/authentication.JWT"
                        }
                    },
                    "400": {
                        "description": "Bad request due to validation errors or malformed input.",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error, something went wrong while processing the request.",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad gateway, likely due to an external service failure.",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            }
        },
        "/v1/auth/sign-in": {
            "post": {
                "description": "Authenticate a user using their email and password. If successful, a JWT is returned which can be used for subsequent authenticated requests.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign-In with email and password",
                "parameters": [
                    {
                        "description": "Request body containing email and password.",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/authentication.SignInParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the generated JWT upon successful authentication.",
                        "schema": {
                            "$ref": "#/definitions/authentication.JWT"
                        }
                    },
                    "400": {
                        "description": "Bad request due to validation errors or malformed input.",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error, something went wrong while processing the request.",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad gateway, likely due to an external service failure.",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "authentication.JWT": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eysjs9a..."
                },
                "expiresIn": {
                    "type": "integer",
                    "example": 300
                },
                "idToken": {
                    "type": "string",
                    "example": "eysjg54g3ba..."
                },
                "notBeforePolicy": {
                    "type": "integer",
                    "example": 150
                },
                "refreshExpiresIn": {
                    "type": "integer",
                    "example": 300
                },
                "refreshToken": {
                    "type": "string",
                    "example": "eynmiks.ewij..."
                },
                "scope": {
                    "type": "string",
                    "example": "email client user ..."
                },
                "sessionState": {
                    "type": "string",
                    "example": "634g43y3..."
                },
                "tokenType": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
        "authentication.RefreshParams": {
            "type": "object",
            "required": [
                "refreshToken"
            ],
            "properties": {
                "refreshToken": {
                    "type": "string",
                    "example": "eyajbhhkd....."
                }
            }
        },
        "authentication.SignInParams": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "*********"
                }
            }
        },
...
        "xerrors.CustomError": {
            "type": "object",
            "properties": {
                "debug": {},
                "details": {},
                "error": {
                    "type": "string"
                },
                "traceID": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}