# internal/api/controller/auth_controller.go
package controller

import (
	"net/http"

	"github.com/gin-gonic/gin"

	validatorutils "github.com/oprimogus/cardapiogo/internal/api/validator"
	"github.com/oprimogus/cardapiogo/internal/core/authentication"
	xerrors "github.com/oprimogus/cardapiogo/internal/errors"
	logger "github.com/oprimogus/cardapiogo/pkg/log"
)

type AuthController struct {
	validator      *validatorutils.Validator
	useCaseSignIn  authentication.UseCaseSignIn
	useCaseRefresh authentication.UseCaseRefresh
}

func NewAuthController(validator *validatorutils.Validator, repository authentication.Repository) *AuthController {
	return &AuthController{
		validator:      validator,
		useCaseSignIn:  authentication.NewUseCaseSignIn(repository),
		useCaseRefresh: authentication.NewUseCaseRefresh(repository),
	}
}

// SignIn godoc
//
//	@Summary		Sign-In with email and password
//	@Description	Authenticate a user using their email and password. If successful, a JWT is returned which can be used for subsequent authenticated requests.
//	@Tags			Authentication
//	@Accept			json
//	@Produce		json
//	@Param			request	body		authentication.SignInParams	false	"Request body containing email and password."
//	@Success		200		{object}	authentication.JWT		"Returns the generated JWT upon successful authentication."
//	@Failure		401		{object}	xerrors.CustomError	"Unauthorized"
//	@Failure		400		{object}	xerrors.CustomError	"Bad request due to validation errors or malformed input."
//	@Failure		500		{object}	xerrors.CustomError	"Internal server error, something went wrong while processing the request."
//	@Failure		502		{object}	xerrors.CustomError	"Bad gateway, likely due to an external service failure."
//	@Router			/v1/auth/sign-in [post]
func (c *AuthController) SignIn(ctx *gin.Context) {
	traceID := ctx.GetString(string(logger.TraceIDKey))
	var params authentication.SignInParams
	err := ctx.BindJSON(&params)
	if err != nil {
		xerror := xerrors.HandleError(err, traceID)
		ctx.JSON(xerror.Status, xerror)
		return
	}
	err = c.validator.Validate(traceID, params)
	if err != nil {
		xerror := xerrors.HandleError(err, traceID)
		ctx.JSON(xerror.Status, xerror)
		return
	}
	jwt, err := c.useCaseSignIn.Execute(ctx, params.Email, params.Password)
	if err != nil {
		xerror := xerrors.HandleError(err, traceID)
		ctx.JSON(xerror.Status, xerror)
		return
	}

	ctx.JSON(http.StatusOK, jwt)
}

// RefreshUserToken godoc
//
//	@Summary		Refresh JWT using a refresh token
//	@Description	Refresh an expired access token by providing a valid refresh token. Returns a new JWT if the refresh is successful.
//	@Tags			Authentication
//	@Accept			json
//	@Produce		json
//	@Param			request	body		authentication.RefreshParams	true	"Request body containing the refresh token."
//	@Success		200		{object}	authentication.JWT		"Returns a new JWT upon successful token refresh."
//	@Failure		400		{object}	xerrors.CustomError	"Bad request due to validation errors or malformed input."
//	@Failure		500		{object}	xerrors.CustomError	"Internal server error, something went wrong while processing the request."
//	@Failure		502		{object}	xerrors.CustomError	"Bad gateway, likely due to an external service failure."
//	@Router			/v1/auth/refresh [post]
func (c *AuthController) RefreshUserToken(ctx *gin.Context) {
	traceID := ctx.GetString(string(logger.TraceIDKey))
	var params authentication.RefreshParams
	err := ctx.BindJSON(&params)
	if err != nil {
		xerror := xerrors.HandleError(err, traceID)
		ctx.JSON(xerror.Status, xerror)
		return
	}
	err = c.validator.Validate(traceID, params)
	if err != nil {
		xerror := xerrors.HandleError(err, traceID)
		ctx.JSON(xerror.Status, xerror)
		return
	}
	jwt, err := c.useCaseRefresh.Execute(ctx, params.RefreshToken)
	if err != nil {
		xerror := xerrors.HandleError(err, traceID)
		ctx.JSON(xerror.Status, xerror)
		return
	}

	ctx.JSON(http.StatusOK, jwt)
}